+++ R/Spectra.R
@@ -1,5 +1,5 @@
-#'@include PlainTextParam.R
-#'@title Methods to save and load contents of a Spectra object
+#' @include PlainTextParam.R
+#' @title Methods to save and load contents of a Spectra object
 #'
 #' @author Philippine Louail
 #'
@@ -41,23 +41,21 @@ setMethod("readMsObject", signature(object = "Spectra",
               var_names <- sub(" =.*", "", fls)
               var_values <- sub(".* = ", "", fls)
               variables <- setNames(var_values, var_names)
+              message("backend ", variables[["backend"]])
               if (!existsMethod("readMsObject", c(variables[["backend"]],
                                                  "PlainTextParam")))
                   stop("Can not read a 'Spectra' object with backend '",
                        variables["backend"], "'")
-              ## Check if the library to load the backend class is available.
-              ## This should also enable backends that are defined in other
-              ## packages than Spectra. Accessing directly the "globalenv" to
-              ## ensure we can access functions/classes there.
-              fun <- getFunction(variables[["backend"]], mustFind = FALSE,
-                                 where = topenv(globalenv()))
-              if (!length(fun))
-                  stop("Can not create an instance of the MsBackend class \"",
-                       variables[["backend"]], "\". Please first load the ",
-                       "library that provides this class and try again.",
+              tryCatch({
+                  object@backend <- readMsObject(
+                      do.call(variables[["backend"]], list()), param, ...)
+              }, error = function(e) {
+                  stop("Failed to load a backend of type '",
+                       variables[["backend"]],
+                       "'. Please load the package defining this type",
+                       " of class and try again.\n - ", e$message,
                        call. = FALSE)
-              b <- readMsObject(fun(), param, ...)
-              object@backend <- b
+              })
               object@processingQueueVariables <- unlist(
                   strsplit(variables[["processingQueueVariables"]],
                            "|", fixed = TRUE))
@@ -101,9 +99,8 @@ setMethod("readMsObject", signature(object = "Spectra",
 #' @noRd
 .export_spectra_processing_queue <- function(x, path = character()) {
     pq <- x@processingQueue
-    if (length(pq))
-        write_json(serializeJSON(pq),
-                   file.path(path, "spectra_processing_queue.json"))
+    write_json(serializeJSON(pq),
+               file.path(path, "spectra_processing_queue.json"))
 }
 
 #' @noRd
@@ -111,3 +108,79 @@ setMethod("readMsObject", signature(object = "Spectra",
     x@processingQueue <- unserializeJSON(read_json(file)[[1L]])
     x
 }
+
+################################################################################
+##
+## alabaster saveObject/readObject
+##
+################################################################################
+#' @rdname AlabasterParam
+setMethod("saveObject", "Spectra", function(x, path, ...) {
+    if (!existsMethod("saveObject", class(x@backend)[1L]))
+        stop("No method to save a backend of type \"", class(x@backend)[1L],
+             "\" available yet")
+    dir.create(path = path, recursive = TRUE, showWarnings = FALSE)
+    saveObjectFile(path, "spectra",
+                   list(spectra =list(version = "1.0")))
+    tryCatch({
+        do.call(altSaveObject,
+                list(x = x@backend, path = file.path(path, "backend")))
+    }, error = function(e) {
+        stop("failed to save 'backend' of ", class(x)[1L], "\n - ",
+             e$message, call. = FALSE)
+    })
+    .export_spectra_processing_queue(x, path = path)
+    altSaveObject(x@processingQueueVariables,
+                  path = file.path(path, "processing_queue_variables"))
+    altSaveObject(x@processing, path = file.path(path, "processing"))
+    altSaveObject(x@metadata, path = file.path(path, "metadata"))
+    altSaveObject(x@processingChunkSize,
+                  path = file.path(path, "processing_chunk_size"))
+})
+
+validateAlabasterSpectra <- function(path = character(),
+                                     metadata = list()) {
+    .check_directory_content(path, c("backend", "processing_queue_variables",
+                                     "spectra_processing_queue.json",
+                                     "processing", "metadata",
+                                     "processing_chunk_size"))
+}
+
+readAlabasterSpectra <- function(path = character(), metadata = list(),
+                                 ...) {
+    if (!requireNamespace("Spectra", quietly = TRUE))
+        stop("Required package 'Spectra' missing. Please install ",
+             "and try again.", call. = FALSE)
+
+    validateAlabasterSpectra(path, metadata)
+    s <- Spectra::Spectra()
+    s@backend <- altReadObject(file.path(path, "backend"), ...)
+    s <- .import_spectra_processing_queue(
+        s, file.path(path, "spectra_processing_queue.json"))
+    s@processingQueueVariables <- altReadObject(file.path(
+        path, "processing_queue_variables"))
+    s@processing <- altReadObject(file.path(path, "processing"))
+    s@metadata <- altReadObject(file.path(path, "metadata"))
+    s@processingChunkSize <- altReadObject(
+        file.path(path, "processing_chunk_size"))
+    validObject(s)
+    s
+}
+
+#' @rdname AlabasterParam
+setMethod("saveMsObject", signature(object = "Spectra",
+                                    param = "AlabasterParam"),
+          function(object, param) {
+              if (file.exists(param@path))
+                  stop("Overwriting or saving to an existing directory is not",
+                       " supported. Please remove the directory defined with",
+                       " parameter `path` first.")
+              saveObject(object, param@path)
+          })
+
+#' @rdname AlabasterParam
+setMethod("readMsObject", signature(object = "Spectra",
+                                    param = "AlabasterParam"),
+          function(object, param, ...) {
+              readAlabasterSpectra(path = param@path, ...)
+          })