+++ NEWS.md
@@ -1,5 +1,16 @@
 # Version 0.0
 
+## Changes in 0.0.4
+
+- Add *alabaster* `saveObject()` and `readObject()` methods as well as
+  `saveMsObject()` and `readMsObject()` methods with `AlabasterParam` param for
+  `MsBackendMzR`, `Spectra`, `MsExperiment` and `XcmsExperiment` objects.
+
+## Changes in 0.0.3
+
+- Implement `saveMsObject()` and `readMsObject()` with `PlainTextParam` for
+  `MsBackendMzR`, `Spectra`, `MsExperiment` and `XcmsExperiment`.
+
 ## Changes in 0.0.2
 
 - Refactor code for text-based import and export of MS data objects.